Refactor the public data pull script
Refactor the public data pull script to make it easier to use from outside the
script, like from an ipython notebook

Testing done:

```
C02KT61MFFT0:e-mission-server shankari$ ./e-mission-py.bash bin/public/request_public_data.py 2016-03-29-12 2016-03-29-14 https://e-mission.eecs.berkeley.edu
Pulling data from 2016-03-29-12 to 2016-03-29-14 (local time)
...
Data loaded to local server!
```
